+++ include/sort.hpp
@@ -88,7 +88,7 @@ void radix(std::vector<int>& arr) {
 }
 
 void insertion(std::vector<int>& arr) {
-    for (size_t i = 1; i < arr.size(); ++i) {
+    for (int i = 1; i < arr.size(); ++i) {
         int key = arr[i];
         int j = i - 1;
         while (j >= 0 && arr[j] > key) {
